+++ src/main/java/net/frozenblock/freezeframe/component/filter/SpecialFilmFilter.java
@@ -15,7 +15,7 @@
  * along with this program; if not, see <https://github.com/FrozenBlock/Licenses>.
  */
 
-package net.frozenblock.freezeframe.item.filter;
+package net.frozenblock.freezeframe.component.filter;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;