+++ src/main/java/net/frozenblock/freezeframe/networking/FFNetworking.java
@@ -22,8 +22,6 @@ import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.frozenblock.freezeframe.networking.packet.CameraTakeScreenshotPacket;
 import net.frozenblock.freezeframe.networking.packet.ChangeItemStackSizePacket;
 import net.frozenblock.freezeframe.networking.packet.ChangeScopeZoomPacket;
-import net.frozenblock.freezeframe.networking.packet.CloneItemStackPacket;
-import net.frozenblock.freezeframe.networking.packet.DeleteItemStackPacket;
 import net.frozenblock.freezeframe.networking.packet.DevelopingTableSyncSelectPhotographIndexPacket;
 import net.frozenblock.freezeframe.networking.packet.OpenBookPagePhotographInventoryPacket;
 import net.frozenblock.freezeframe.networking.packet.OpenFilmScreenPacket;
@@ -68,12 +66,6 @@ public class FFNetworking {
 
 		c2sRegistry.register(OpenBookPagePhotographInventoryPacket.PACKET_TYPE, OpenBookPagePhotographInventoryPacket.CODEC);
 		ServerPlayNetworking.registerGlobalReceiver(OpenBookPagePhotographInventoryPacket.PACKET_TYPE, OpenBookPagePhotographInventoryPacket::handle);
-
-		c2sRegistry.register(DeleteItemStackPacket.PACKET_TYPE, DeleteItemStackPacket.CODEC);
-		ServerPlayNetworking.registerGlobalReceiver(DeleteItemStackPacket.PACKET_TYPE, DeleteItemStackPacket::handle);
-
-		c2sRegistry.register(CloneItemStackPacket.PACKET_TYPE, CloneItemStackPacket.CODEC);
-		ServerPlayNetworking.registerGlobalReceiver(CloneItemStackPacket.PACKET_TYPE, CloneItemStackPacket::handle);
 
 		c2sRegistry.register(ChangeItemStackSizePacket.PACKET_TYPE, ChangeItemStackSizePacket.CODEC);
 		ServerPlayNetworking.registerGlobalReceiver(ChangeItemStackSizePacket.PACKET_TYPE, ChangeItemStackSizePacket::handle);